+++ src/syncio.c
@@ -153,7 +153,7 @@ int fwriteBulkObject(FILE *fp, robj *obj) {
     if (obj->encoding == REDIS_ENCODING_INT) {
         return fwriteBulkLongLong(fp,(long)obj->ptr);
     } else if (obj->encoding == REDIS_ENCODING_RAW) {
-        return fwriteBulkString(fp,obj->ptr,sdslen(obj->ptr));
+        return fwriteBulkString(fp,obj->ptr,(unsigned long)sdslen(obj->ptr));
     } else {
         redisPanic("Unknown string encoding");
     }
